Fixed legacy imports

+++ server/internal/model/changelog.go
@@ -3,8 +3,8 @@ package model
 type Changelog struct {
 	Timestamp string  `json:"timestamp"`
 	Action    string  `json:"action"`
-	From      *int32  `json:"from"`
-	To        *int32  `json:"to"`
+	From      *int16  `json:"from"`
+	To        *int16  `json:"to"`
 	Reason    *string `json:"reason"`
 	Level     Level   `json:"level"`
 	List      []Level `json:"list"`
